refactor : 사이드 navigationbar 컴포넌트로 분리 적용
