+++ src/Routes/Tweet.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Tweet() {
+  return <div>Tweet</div>;
+}
+
+export default Tweet;
